backend/pyproject.toml
[project]
name = "intervue-backend"
version = "0.1.0"
description = "Resume-aware AI behavioral interview platform"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "openai>=1.50.0",
    "pdfplumber>=0.11.0",
    "httpx>=0.27.0",
    "authlib>=1.3.0",
    "itsdangerous>=2.2.0",
    "python-multipart>=0.0.9",
    "redis[hiredis]>=5.0.0",
    "pydantic-settings>=2.5.0",
    "python-jose[cryptography]>=3.3.0",
    "bcrypt>=4.2.0",
]

[tool.setuptools.packages.find]
where = ["."]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.6.0",
]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
